remove tests notebooks from doc/contents
