+++ test/tests/groovy/javaWidgetsTest.js
@@ -21,7 +21,7 @@ describe('Java widgets notebook', function () {
 
   beforeAll(function () {
     beakerxPO = new BeakerXPageObject();
-    beakerxPO.runNotebookByUrl('/notebooks/doc/contents/JavaWidgets.ipynb');
+    beakerxPO.runNotebookByUrl('/notebooks/test/notebooks/groovy/JavaWidgetsTest.ipynb');
   }, 2);
 
   afterAll(function () {
@@ -35,51 +35,42 @@ describe('Java widgets notebook', function () {
     }, 2);
 
     it('Widget value = 60', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(1, '60');
       var codeCell = beakerxPO.getCodeCellByIndex(0);
       expect(codeCell.$('div.widget-readout').getText()).toBe('60')
     });
 
     it('Widget value = 76', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(2, '76');
       var codeCell = beakerxPO.getCodeCellByIndex(0);
       expect(codeCell.$('div.widget-readout').getText()).toBe('76')
     });
 
     it('Widget description = "desc1"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(3, 'desc1');
       var codeCell = beakerxPO.getCodeCellByIndex(0);
       expect(codeCell.$('label.widget-label').getText()).toBe('desc1')
     });
 
     it('Cell output contains "false"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(4, 'false');
-
     });
 
     it('Cell output contains "50"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(5, '50');
     });
 
     it('Cell output contains "horizontal"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(6, 'horizontal');
       var codeCell = beakerxPO.getCodeCellByIndex(0);
       expect(codeCell.$('div.ui-slider-horizontal').isExisting()).toBeTruthy();
     });
 
     it('Cell output contains "20"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(8, '20');
     });
 
     it('Cell output contains "true"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(9, 'true');
       //TODO visible hasn't worked yet
     });
@@ -92,27 +83,22 @@ describe('Java widgets notebook', function () {
     }, 2);
 
     it('Widget value = 10', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(11, '10');
     });
 
     it('Widget value = 110', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(12, '110');
     });
 
     it('Cell output contains "50"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(13, '50');
     });
 
     it('Cell output contains "20"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(14, '20');
     });
 
     it('Cell output contains "horizontal"', function () {
-      beakerxPO.kernelIdleIcon.waitForEnabled();
       beakerxPO.runCallAndCheckOutputText(15, 'horizontal');
     });
   });
