+++ test/tests/java/javaTest.js
@@ -14,14 +14,14 @@
  *  limitations under the License.
  */
 
-var BeakerXPageObject = require('./beakerx.po.js');
+var BeakerXPageObject = require('../beakerx.po.js');
 var beakerxPO;
 
 describe('JavaTutorial notebook', function () {
 
   beforeAll(function () {
     beakerxPO = new BeakerXPageObject();
-    beakerxPO.runNotebookByUrl('/notebooks/doc/contents/javaTutorial.ipynb');
+    beakerxPO.runNotebookByUrl('/notebooks/test/notebooks/java/JavaTest.ipynb');
   }, 2);
 
   afterAll(function () {
@@ -65,10 +65,10 @@ describe('JavaTutorial notebook', function () {
   });
 
   describe('Run 7th cell. ', function () {
-    it('Output contains "beakerx/doc/contents"', function () {
+    it('Output contains "beakerx/test/notebooks"', function () {
       beakerxPO.runCodeCellByIndex(5);
       beakerxPO.kernelIdleIcon.waitForEnabled();
-      beakerxPO.runCallAndCheckOutputText(6, 'beakerx\.doc\.contents');
+      beakerxPO.runCallAndCheckOutputText(6, 'beakerx\.test\.notebooks');
     }, 2);
   });
