+++ test/tests/scala/scalaTest.js
@@ -21,7 +21,7 @@ describe('Scala notebook', function () {
 
   beforeAll(function () {
     beakerxPO = new BeakerXPageObject();
-    beakerxPO.runNotebookByUrl('/notebooks/doc/contents/scala/scala.ipynb');
+    beakerxPO.runNotebookByUrl('/notebooks/test/notebooks/scala/ScalaTest.ipynb');
   }, 2);
 
   afterAll(function () {
